{"types":"C","runs":27}

+++ Problem21.java
@@ -16,7 +16,7 @@ public class Problem21 extends GraphicsProgram {
 		
 	}
 	private void drawHorizontalLines() {
-		double n = getHeight() / 10;
+		double n = (double)getHeight() / 10;
 		for (int y = 0; y <= getHeight(); y += n) {
 			GLine hLine = new GLine(0, y, getWidth(), y);
 			add(hLine);
